Split pipeline into its own class
